refactor: move logic to operationsHandler
